blog_app/requirements.txt
Django
Pillow
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
PyJWT
factory_boy